trampoline improvements
* improve trampoline by detecting unrecognized named parameters
* move proc probing and trampoline stuff into __init__.py
